Rewritten pds-files source script to account for Win vs. Linux
The way the glob expansion is treated on the command line is
completely different across systems:

In Linux, the shell expands the wildcards (incl. recursive wildcards
for modern shells!), then the expanded array gets passed to the
spawned process as argv

In Windows, the glob argument is passed to the spawned process as it
is, and it the responsibility of the process to do the expansion.